git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/branches/issue-clo-1593@15069 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/util/file/FileUtils.java
@@ -292,11 +292,17 @@ public class FileUtils {
     		fileURL = fileURL.substring((FILE_PROTOCOL+":").length());
     	}
 
+        final String protocol = getProtocol(fileURL);
+        if (DICTIONARY_PROTOCOL.equalsIgnoreCase(protocol) || PORT_PROTOCOL.equalsIgnoreCase(protocol)) {
+            return new URL(contextURL, fileURL, GENERIC_HANDLER);
+        }
+
     	 //first we try the custom path resolvers
     	for (CustomPathResolver customPathResolver : customPathResolvers) {
-    		try{
+    		// CLO-978 - call handlesURL(), don't catch any MalformedURLExceptions
+    		if (customPathResolver.handlesURL(contextURL, fileURL)) {
     			return customPathResolver.getURL(contextURL, fileURL);
-    		}catch(MalformedURLException ex) {}
+    		}
     	}
     	
     	// standard url
@@ -335,15 +341,12 @@ public class FileUtils {
 			return new URL(null, type.getId() + ":(" + archiveFileUrl.toString() + ")#" + anchor, new ArchiveURLStreamHandler(contextURL));
 		}
 		
-		String protocol = getProtocol(fileURL);
-		if (DICTIONARY_PROTOCOL.equalsIgnoreCase(protocol) || PORT_PROTOCOL.equalsIgnoreCase(protocol)) {
-			return new URL(contextURL, fileURL, GENERIC_HANDLER);
-		} else if (!StringUtils.isEmpty(protocol)) {
-			// unknown protocol will throw an exception,
-			// standard Java protocols will be ignored;
-			// all Clover-specific protocols must be checked before this call
-			new URL(fileURL);
-		}
+		if (!StringUtils.isEmpty(protocol)) {
+            // unknown protocol will throw an exception,
+            // standard Java protocols will be ignored;
+            // all Clover-specific protocols must be checked before this call
+            new URL(fileURL);
+        }
 		
 		if (StringUtils.isEmpty(protocol)) {
 			// file url
@@ -2307,15 +2310,14 @@ public class FileUtils {
 		if (Thread.currentThread().isInterrupted()) {
 			throw new IOException("Interrupted");
 		}
-		if (root != null && root.exists()) {
-			if (root.isDirectory()) {
+		if (root != null && root.isDirectory()) {
 				File[] children = root.listFiles();
 				if (children != null) {
 					for (int i = 0; i < children.length; i++) {
 						deleteRecursively(children[i]);
 					}
 				}
-			}
+
 			return root.delete();
 		}
 		return false;
